move discard to actions
