+++ hundredandten/action.py
@@ -2,6 +2,7 @@
 from dataclasses import dataclass, field
 
 from hundredandten.constants import BidAmount
+from hundredandten.deck import Card
 
 
 @dataclass(order=True)
@@ -12,3 +13,10 @@ class Bid:
 
     def __bool__(self) -> bool:
         return self.amount.value > 0
+
+
+@dataclass
+class Discard:
+    '''A class to keep track of one player's discard action'''
+    identifier: str
+    cards: list[Card]
