+++ hundredandten/game.py
@@ -10,7 +10,7 @@ from hundredandten.constants import (HAND_SIZE, WINNING_SCORE, Accessibility,
                                      GameStatus, RoundRole, RoundStatus,
                                      SelectableSuit)
 from hundredandten.deck import Deck
-from hundredandten.discard import Discard
+from hundredandten.action import Discard
 from hundredandten.group import Group, Person, Player
 from hundredandten.hundred_and_ten_error import HundredAndTenError
 from hundredandten.round import Round
