+++ tests/game/test_discard.py
@@ -1,8 +1,8 @@
 '''Test behavior of the Game when going around discarding'''
 from unittest import TestCase
 
+from hundredandten.action import Discard
 from hundredandten.constants import HAND_SIZE, RoundStatus
-from hundredandten.discard import Discard
 from hundredandten.hundred_and_ten_error import HundredAndTenError
 from tests import arrange
